+++ packages/api/src/submittable/Result.ts
@@ -16,15 +16,19 @@ export default class SubmittableResult implements SubmittableResultImpl {
   }
 
   public get isCompleted (): boolean {
-    return this.isError || this.isFinalized;
+    return this.isError || this.status.isInBlock || this.status.isFinalized;
   }
 
   public get isError (): boolean {
-    return this.status.isDropped || this.status.isInvalid || this.status.isUsurped;
+    return this.status.isDropped || this.status.isFinalityTimeout || this.status.isInvalid || this.status.isRetracted || this.status.isUsurped;
   }
 
   public get isFinalized (): boolean {
     return this.status.isFinalized;
+  }
+
+  public get isInBlock (): boolean {
+    return this.status.isInBlock;
   }
 
   /**